requirements.txt
flask
torch
transformers
pillow
numpy
requests